[Test] Compare exp3.json against the managed projection
The third Cubism document, and the one that settles the number-format
question. Physics puts every float through .NET's "0.###", motion uses that
for segment lists and Newtonsoft's default float for everything else, and
the expression document is serialized with no converter at all, so it is
default float throughout.

With all three in the differential that distinction is held down rather than
asserted: getting it backwards on any one of them fails that document while
the others pass. On its own each looked like a formatting detail; together
they are three different behaviours in the same extractor.

The defect here was the f64 widening again, the same one physics and motion
had. Verified in reverse -- putting serde_json's f64 output back fails the
comparison -- because the fix went in before the test ran, and a test that
has never failed has not been shown to work.

pose3 and cdi3 stay out. Neither is one behaviour in and one document out:
the managed side assembles them by walking a model's parts and parameters,
so they need a fixture with GameObject and Transform wiring. Neither carries
a float field either, which is where every defect in this area has been.

+++ crates/assetstudio-python/src/lib.rs
@@ -2829,7 +2829,7 @@ impl PyAssetStudio {
                 py,
                 PyCubismExpressionParameter {
                     id: parameter.id,
-                    value: parameter.value,
+                    value: f64::from(parameter.value),
                     blend: cubism_expression_blend_name(parameter.blend).to_owned(),
                 },
             )?);
@@ -2838,8 +2838,8 @@ impl PyAssetStudio {
             path_id: expression.path_id,
             source_name: expression.source_name,
             expression_type: expression.expression_type,
-            fade_in_time: expression.fade_in_time,
-            fade_out_time: expression.fade_out_time,
+            fade_in_time: f64::from(expression.fade_in_time),
+            fade_out_time: f64::from(expression.fade_out_time),
             parameters,
             json,
         })
